update static links to static.slabs.io
relates to #220 in slabs core

For now I've updated the static links that load the selector interface into
the page to link to static.slabs.io. There is further work to implement
a proper deployment chain for the client work. This will require a mechanism
in the pagescrape slab to link to the correct version. For now at least we have
a working interface.