fix bug where multiple items are selected
refers to #220 in slabs-core

This commit fixes a bug where 2 items are selected but they do not belong in a
group. NodeLists, JQuery (and CheerioJS) return Array-like structures, so you
can't treat them as actual Arrays. The 'fix' is to Array.prototype.slice your
ALS and then perform the operation on the returned array.